Hide JSON output flag (#1209)
The JSON format is unstable and intended for internal use, so we don't
want users to get impacted by depending on a format that changes out
from under them.

We are tracking a feature proposal to add an output format for the SARIF
standard [here](https://github.com/orgs/qltysh/discussions/1205), which
will enable a stable machine-readable output format.