Pre-commit (fmt) and pre-push (check) git hooks (#1243)
Note: A `git push` operation can push one or more references, and the
references that are pushed may not be a reference that is checked out in
the local working copy.

Since the `check` command always works on the files that are checked out
in the working directory, the files that we check may not match the
actual files being pushed.

Also, it's unclear what the correct behavior should be if multiple
references are being pushed simultaneously. Mt first gut reaction is
probably to no-op, since we don't want to disrupt the user.

Git passes a pre-push hook data on stdin which contains the commit ID of
the local and the remote, and this PR uses the remote commit ID as our
"upstream", so that the `check` execution will target the files that are
changed in the diff.


Also -- I am shelling out to `git add` and it would be nice to replace
this with the equivalent behavior emulated by the git2 crate.