Bundle plugin definitions into compiled binary (#1262)
Opening a POC for feedback...

This approach uses rust-embed to build the TOML files directly into the
crate. Therefore, we could support a third type of Source, in addition
to Git-based and local sources: A "default" source which is compiled
into the CLI.

So instead of generating this:

```toml
[[source]]
name = "default"
repository = "https://github.com/qltysh/qlty"
tag = "v0.450.0"
```

We could generate something like this:

```toml
[[source]]
name = "default"
default = true
```

This would mean that for default configurations the CLI version _is_ the
plugins version.

---------

Co-authored-by: qltysh[bot] <168846912+qltysh[bot]@users.noreply.github.com>