Bump default Hadolint to 2.12.1-beta and auto-enable (#1299)
The most recent hadolint release has known segfaults on MacOS, which
appear to be patched in this pre-release.

This PR contains three changes:

1. Bump Hadolint to 2.12.1-beta
2. Re-enable Hadolint tests on MacOS
3. Re-enable Hadolint initialization when Dockerfiles are found
4. When validating installed plugin versions, apply the regex against
the _declared_ plugin version in addition to the version output

The third change is necessary because we need to declare `version =
"2.12.1-beta"` and `hadolint --version` in that case will print only
`2.12.1`. Since `"2.12.1" != "2.12.1-beta"` the tool validation would
fail.

If we use the version regex `(\d+\.\d+\.\d+)` to capture the version
from `"2.12.1-beta"` before we compare, we see that `"2.12.1" ==
"2.12.1"` and installation validation can succeed.