Set plugin `mode=disabled`(instead of removing whole plugin) when running `qlty plugins disable PLUGIN` (#1315)
- Changes behavior of the `qlty plugins disable PLUGIN` to set the
plugin mode to `disabled` (instead of removing it from the qlty.toml)
- Adds a couple error checks: errors if no plugins in qlty.toml or if
the plugin referenced in the command is not found in their qlty.toml
(used `upgrade` command as reference)

Couple notes:

1. The conversion of `IssueMode` to a string duplicates the
serialization macro strings (`#[serde(rename = "block")]` e.g.) defined
in the struct itself. From what I could tell, if we want to avoid this
duplication, we could either a) use `serde_json` to translate the
IssueMode value to JSON, and then extract the string out of the JSON
(yuck). Or, instead, remove these macros and define our own
serialization, using the new `to_string` method defined. I'm less sure
where the serialization is actually happening, so not confident I could
do this.
2. I looked/tested how `qlty plugins enable PLUGIN` behaves. If a plugin
is disabled (with the `disable` command or by hand), you get an error
message "plugin PLUGIN already enabled". So we'll have to change this
behavior to just remove the disabled mode line (I'm guessing).