Ability to abort qlty check during Git hooks (#1306)
- If `qlty check` is being run as a git hook
- Allow the user to press `<enter>` to short-circuit execution
- This is useful both for speed when needed and also to skip past
otherwise-blocking issues

Note: Git does not use a tty for it's hooks. Therefore, reading from
stdin as usual does not work, so instead we manually try to open and
read `/dev/tty`. (Unsure what if anything we would need to do to make
this work on Windows.)

### Also

I worked on this on an airplane with no WiFi and so there's some
additional teaks in here:

1. Improve qlty configuration, including getting ESLint working on new
plugins folder structure
2. Reduce output when not running `--verbose`
3. Fix some minor output errors
4. Fix a panic in `qlty fmt` with a non-existent path argument