enhance `qlty plugins enable PLUGIN` to support enabling disabled plugins (#1316)
Previously `qlty plugins disable PLUGIN` disabled a plugin by entirely
removing the plugin definition from the `qlty.toml` file. In this world
`qlty plugins enable PLUGIN` worked regardless of whether the plugin was
never configured to begin with or was configured but removed with `qlty
plugins disable PLUGIN`.

This changed in #1315 when we changed the behavior of `qlty plugins
disable PLUGIN` to instead mark the plugin as disabled (`mode =
"disabled"`). Now, a `disabled` plugin differs from one never configured
to begin with. Because the `enable` subcommand was not "mode-aware" --
it assumed if the plugin was defined in the `qlty.toml`, regardless of
its "mode", that it was enabled. This had the unfortunate user
experience of informing a user that their plugin was already enabled
when in fact it was explicitly disabled.

This PR makes the `enable` subcommand mode-aware: if a plugin is
explicitly disabled with `mode = "disabled"` this PR removes that line
from their configuration, effectively re-enabling that plugin.

This does mean that if at one point the plugin was configured with a
mode other than the default (currently `block`), this would need to be
put back into place.