Update CLI runners (#1386)
Run times for CLI/plugin tests on macOS and Ubuntu (Windows not
changed-- failures are unrelated):

CLI (cli.yml):

*
[ubuntu-latest-16-cores](https://github.com/qltysh/qlty/actions/runs/12588323414/job/35085973505#logs):
1m44s
*
[ubuntu-latest](https://github.com/qltysh/qlty/actions/runs/12588501607/job/35086497825#logs):
1m42s
*
[macos-15-xlarge](https://github.com/qltysh/qlty/actions/runs/12588323414/job/35085973271#logs):
3m42s
*
[macos-15](https://github.com/qltysh/qlty/actions/runs/12588501607/job/35086497585#logs):
1m41s

Plugins (plugins.yml):

*
[ubuntu-latest-16-cores](https://github.com/qltysh/qlty/actions/runs/12588323409/job/35085973461#logs):
5m45s
*
[ubuntu-latest](https://github.com/qltysh/qlty/actions/runs/12588501606/job/35087704835#logs):
6m45s
*
[macos-15-xlarge](https://github.com/qltysh/qlty/actions/runs/12588323409/job/35085973262#logs):
7m40s
*
[macos-15](https://github.com/qltysh/qlty/actions/runs/12588501606/job/35087705326#logs):
5m50s

Seems like in both cases, running the CI using standard runners is
either equally fast, or, in some cases, slightly faster or slower(*). I
think the recommendation should be to switch to standard runners.

(*): this is not averaged across runs, it's possible we're exhibiting
best-case and worst-case performance for the sample set. The comparison
is not meant to be exhaustive since the stddev across runs can be
extremely high in real-world usage. Seeing any form of reasonable <2min
overlap should be enough to consider these timings functionally
equivalent.