Update CI runners for release builds (#1387)
Related to #1386

Not sure if we explicitly need an arm64 host to build an arm64 linux  builds since Rust should be able to cross-compile across architectures (we explicitly already do this on macOS). I think it's worth validating this via a release with standard runners and have high confidence it would work.