Add `cache prune` subcommand (#1408)
Uses the file MAC times to prune the results cache, logs, and outfiles.
This is a fast-and-dumb approach rather than trying to analyze
everything and figure out which results cache files are truly no longer
needed.

Some filesystems may not track access times (`atime`) which provides
visibility into cache reads, in which case it will fallback to modified
time (`mtime`).

Also, adds a `plugin_cachedir` to the Library, to keep data cached by
the plugins isolated from data that we cache. (In our cache, the Clippy
plugin ends up with ~2 GB of "cache" which is really compilation
artifacts.) Since we can't be sure of the impact of deleting
some-but-not-all of a plugin's cached data, we delete it all.

### Future improvements

- Enhance `cache prune` to have an argument for whether to clear the
plugin caches or not, since that data is likely to be large but also
slow to re-generate
- Automatically run `cache prune` logic regularly _without_ clearing the
plugin caches.