Fix panic when path for `qlty check <path>` is outside qlty directory (#1426)
The walker attempts to strip the "root qlty directory" off of each path
when iterating over args, but if the prefix does not match, the
offending line unwraps a None option and panics.

This fixes the panic, although in general, the behavior for this
situation is undefined (qlty does not support calling `check <path>` on
paths that are not in the root qlty directory).

Ideally all path arguments should be validated and an error should be
printed when an invalid path is found, but let's start by removing
panics.