Skip formatters when sampling issues during init (#1474)
Right now we also run formatters during `qlty init` if user approves of
sampling issues, which then prompts user to format files which can
happen in a loop and often be confusing during initialization for a
user.

I think the only way to avoid that is to skip running formatters at this
stage.