Change default behavior for missing output (#1466)
Prior to this PR, if a linter exited with a success exit code but did
not produce any output, we would hand an empty string to the parser. In
substantially all situations, this will result in a confusing EOF error
from the parser, because it will try to e.g. parse an empty string as
JSON.

This PR changes the default -- If a linter exits with a success exit
code, but does not produce output, we will treat that as a Lint Error
and skip the parser phase entirely.

This behavior can be overridden by two alternative behaviors: `parse` or
`no_issues`. `parse` will invoke the parser with an empty string, and
`no_issues` will treat the empty output as if it produced output along
the lines of `issues = []`

(Previously, we had a `missing_output_as_error` boolean. The new
`output_missing` enum replaces that.)