Fix path target mode when used in subdirectory  (#1476)
Right now if you apply a formatting change bubbled up from `qlty check`
command while in a sub directory, the application fails with an
`Result::unwrap() on an Err value`.

The root of that issue lies in the fact that when you apply the format
this way, formatter is run with only selected file paths (which are
project root relative and not relative to cwd) and TargetMode is set to
Path when that happens. Instead of normalizing relative paths based on
cwd I think just using absoluete paths is a better approach and less
likely to cause any issues.