Treat missing phpstan output as error (#1504)
In some cases, phpstan will exit with code 1. From the docs, it is a bit
unclear, but I believe that phpstan uses exit code 1 also to mean
"issues found".

The problem is that if phpstan crashes, it exits with code 1, but does
not produce any output on stdout. Prior to this PR, we would consider
this to be a _successful_ lint invocation and then hand an empty string
to the parser which would raise an EOF error. After this PR, if phpstan
produces no stdout, we consider that to be a _lint_ phase error and we
will not move to the parser.

In terms of outcome, it is the same -- Either behavior results in an
error. However, in the case of a Lint Error, we will print the warnings
and errors from stderr to the terminal, whereas in the case of a Parse
Error we do not. So specifying the `output_missing = "error"` behavior
in this case increases that clarity of phpstan errors for customers.