Move AI autofix logic into transformers so suggestions can be cached (#1519)
My original attempt at this implementation was manually caching the
autofix() step, but it seemed very messy (that impl is not provided in
this PR).

This approach feels a lot cleaner because it utilizes IssueTransformer
to mutate the issue. We can also hook directly into progress to provide
a much more understandable inline UX.