Fix GitDiff calculation of repository paths (#1550)
* GitDiff::compute path is passed as the root of repository, but we use
the .parent() when computing diffs
* Our test `sample_repo()` returns a Repository object whose `.path()`
returns the subdir `.git/` directory, meaning we are testing the
function differently from how it is used.

tldr: the parent() is unnecessary. This causes matching issues when
running from subdirectories (issues are filtered out because they do not
match the path). To be honest, I'm not 100% sure why this works when
running from root, but we have been passing the wrong path for those
files too (i.e. for `/path/to/qltysh/qlty` we used `/path/to/qltysh` as
the repo dir).