Copy lockfiles when no package_filters (#1658)
A common issue I have noticed in testing this is that if the version of
the linter in the lockfile is specified to be different than in
qlty.toml (often the default known_good_version) we get the version
mismatch error during version check post tool install.

Maybe we should skip version check in this case?