Fix panic: byte index 80 is not a char boundary (#1955)
The stack trace is only partially helpful, potentially due to
concurrency in this part of the code.

The panic error is `byte index 80 is not a char boundary` and my strong
suspicion is that it comes from this line:

```rust
format!("{}...", &issue.message[..80])
```

The input `80` to the slice operator works is bytes not characters. If
that byte location is not a character boundary, it panics.

Therefore, any unprotected string slice where the string is arbitrary
content that may include multi-byte characters is a panic bug.