Fix CI: Update flake8 snapshots (#2002)
These snapshots became out-of-date once we fixed the bug to copy linter
configs, because the `.flake8` config file suddenly started to be copied
(before, it was ignored).

I confirmed that flake8 passes the tests in main if the `.flake8` file
is removed, so it's working fine.

This is a blind update of the snapshots to get CI working